refactor: extraer DTOs de firebase, geocoding y schedule_exception a archivos separados
- Crear firebase.go, geocoding.go, schedule_exception.go
- Seguir patrón existente: DTOs en *.go, lógica en *_controller.go
- 19/19 controllers ahora cumplen el patrón de separación
