+++ handlers/firebase.go
@@ -1,0 +1,6 @@
+package handlers
+
+// FirebaseTokenResponse represents the response for GET /auth/firebase-token
+type FirebaseTokenResponse struct {
+	FirebaseToken string `json:"firebase_token"`
+}
